Güncelleme , silme, canlı  user kontrolü ve yetkilenddirmesi ve sitedeki bazı katmanlar düzeltildi büyük çaplı değişiklikler yapıldı
