+++ Controllers/HomeController.cs
@@ -27,10 +27,6 @@ namespace MoodApp.Controllers
         {
             return View();
         }        
-        public IActionResult akif()
-        {
-            return View();
-        }
 
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
         public IActionResult Error()
