+++ Views/Auth/Dashboard.cshtml
@@ -1,5 +1,7 @@
 
 <h2>İşlem Yapacağınız Alan</h2>
 
-<a href="Add">Ekleme Sayfası</a><br>
-<a href="Update">Silme ve Güncelleme Sayfası</a>
+<a href="/Question/Add">Ekleme Sayfası</a><br>
+<a href="/Question/Update">Silme ve Güncelleme Sayfası</a><br>
+<a href="AssignRole">Yetki ataması</a><br>
+<a href="AddUser">Yeni Kullanıcı Ekle</a>
