--- Views/Auth/UpdateQuestion.cshtml
@@ -1,41 +1,0 @@
-<script>
-    function deleteElement(id)
-    {
-        var btnDelete=  document.getElementById("btn"+id).remove();
-        document.getElementById("txtDescription"+id).remove();
-        document.getElementById("txtOption"+id).remove();
-    }
-</script>
-<form method="post" enctype="multipart/form-data" asp-controller="Auth" asp-action="UpdateQuestion">
-
-@{
-if(ViewBag.QuestionList is not null)
-{
-    if(ViewBag.QuestionList.Rows.Count==0)
-    {
-        ViewBag.Message="Böyle bir kayıt yok...";
-    }
-    else
-    {
-        <textarea name="txtQuestion">@ViewBag.QuestionList.Rows[0]["question"]</textarea><br>
-        string optName;
-        string btnName;
-        string dscName;
-        int sayac=1;
-        for (int i = 0; i < ViewBag.QuestionList.Rows.Count; i++)
-        {
-            optName="txtOption"+@sayac;
-            dscName="txtDescription"+@sayac;
-            btnName="btn"+@sayac;
-            <input type="text" name="@optName" id="@optName" placeholder="Seçenek..." value="@ViewBag.QuestionList.Rows[i]["answer"]"><br>
-            <textarea name="@dscName" id="@dscName" placeholder="Açıklama...">@ViewBag.QuestionList.Rows[i]["text"]</textarea><br>
-            <input type="button" id="@btnName" value="*Sil*" onclick="deleteElement(@i)"><hr>
-            sayac++;
-        }
-        <input type="submit" value="Submit"/>
-    }
-}
-<h2>@ViewBag.Message</h2>
-}
-
-</form>
