+++ Views/Question/Add.cshtml
@@ -53,7 +53,7 @@
         <h2>@ViewBag.Message</h2>
     }
 }
-<form method="post" enctype="multipart/form-data" asp-controller="Auth" asp-action="Add">
+<form method="post" enctype="multipart/form-data" asp-controller="Question" asp-action="Add">
             Soru: <br>
             <textarea name="txtQuestion">Soruyu yazınız....</textarea><button type="button" onclick="addInput()" name="btnAdd">*Seçenek Ekle*</button>
             <br>
