+++ Views/Question/Test.cshtml
@@ -1,0 +1,5 @@
+@{
+    Layout=null;
+}
+
+@ViewBag.Message
